Added Support for Boost 1.70+
Boost (at some point) removed the `boost::asio::acceptor::get_io_service()`
method. To fix this, I added the io_service as a parameter for the Router
as a pointer (`boost::shared_ptr`, though I don't know if a Shared Pointer)
is necessary.

In accordance with this I also edited the `app/main.cpp` file.